Move Overlay to ./containers
